+++ ProyectoCine/src/main/java/com/example/demo/controller/FuncionController.java
@@ -20,6 +20,8 @@ import com.example.demo.repository.HorarioRepository;
 import com.example.demo.repository.PeliculaRepository;
 import com.example.demo.repository.SalaRepository;
 
+import jakarta.servlet.http.HttpSession;
+
 @Controller
 public class FuncionController {
 	@Autowired
@@ -32,7 +34,10 @@ public class FuncionController {
 	private HorarioRepository repositoryHorario;
 	
 	@GetMapping("/funciones")
-	public String home(Model model) {
+	public String home(Model model, HttpSession session) {
+		if(session.getAttribute("empleado") == null) {
+			return "redirect:/";
+		}
 		List<Funcion> listaFuncion = repositoryFuncion.findAll();
 		model.addAttribute("listaFuncion",listaFuncion);
 		
@@ -40,7 +45,10 @@ public class FuncionController {
 	}
 	
 	@GetMapping("/registrar_funcion")
-	public String showRegistrarFuncion(Model model) {
+	public String showRegistrarFuncion(Model model, HttpSession session) {
+		if(session.getAttribute("empleado") == null) {
+			return "redirect:/";
+		}
 		model.addAttribute("funcion", new Funcion());
 		
 		List<Pelicula> listaPeliculas = repositoryPelicula.findAll();
@@ -55,13 +63,19 @@ public class FuncionController {
 	}
 	
 	@PostMapping("/registrar_funcion")
-	public String registrarFuncion(@ModelAttribute Funcion funcion, Model model) {
+	public String registrarFuncion(@ModelAttribute Funcion funcion, Model model, HttpSession session) {
+		if(session.getAttribute("empleado") == null) {
+			return "redirect:/";
+		}
 	    repositoryFuncion.save(funcion);
 		return "redirect:/funciones";
 	}
 	
 	@GetMapping("/editar_funcion/{id}")
-	public String mostrarFormularioEdicion(Model model, @PathVariable("id") Integer id) {
+	public String mostrarFormularioEdicion(Model model, @PathVariable("id") Integer id, HttpSession session) {
+		if(session.getAttribute("empleado") == null) {
+			return "redirect:/";
+		}
 	    Funcion funcionActualizar = repositoryFuncion.findById(id).get();
 	    List<Pelicula> listaPeliculas = repositoryPelicula.findAll();
 	    List<Sala> listaSalas = repositorySala.findAll();
@@ -76,18 +90,34 @@ public class FuncionController {
 	}
 	
 	@PostMapping("/editar_funcion")
-	public String actualizarFuncion(@ModelAttribute Funcion funcion, Model model) {
-		
+	public String actualizarFuncion(@ModelAttribute Funcion funcion, Model model, HttpSession session) {
+		if(session.getAttribute("empleado") == null) {
+			return "redirect:/";
+		}
 		repositoryFuncion.save(funcion);
 	    return "redirect:/funciones";
 	}
 	
 	@GetMapping("/delete/{id}")
-	public String eliminarFuncion(@PathVariable("id")Integer id) {
+	public String eliminarFuncion(@PathVariable("id")Integer id, HttpSession session) {
+		if(session.getAttribute("empleado") == null) {
+			return "redirect:/";
+		}
 		repositoryFuncion.deleteById(id);
 		return "redirect:/funciones";
 	}
 	
-	
+	@GetMapping("/escoger_funcion")
+	public String showEscogerFuncion(HttpSession session, Model model) {
+		if(session.getAttribute("empleado") == null) {
+			System.out.println("La sesión se cerró");
+			return "redirect:/";
+		}
+		
+		List<Funcion> funciones = repositoryFuncion.findAll();
+		model.addAttribute("funciones", funciones);		
+											
+		return "escoger_funcion";
+	}
 	
 }
